Separating the routing in its owns module
